backend/requirements.txt
Django==4.2.15
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
dj-database-url==2.2.0
gunicorn==23.0.0
django-cors-headers==4.4.0